Performances now finish correctly.
Performances now end when all the performing
trackWorkers have sent the last moment in their
final trk.
The final trk now completes in accordance
with its inputControls (set in the score) when the
performer sends the final noteOff.
Currently this can mean "stop immediately",
"fade to end of trk" or "complete the trk as
defined in the score".