AssistantPerformer creates its own markers layer
It is no longer necessary for Moritz to create an
invisible, clickable layer. The AP does this itself.
Currently, the markers themselves are still written
to each page of the score, but I am about to create
them inside this application layer instead.
These things should not be part of the score.