Start- running- and end-markers now in markers layer.
It is no longer necessary for Moritz to write these
objects into each score. They are created automatically
as part of this Assistant Performer application, and
added to the group (=layer) containing the transparent,
rectangle that responds to clicks when setting the
marker positions.
This simplifies both Moritz and the editing of score
files in Inkscape.
TODO 1: Use switches here when reading score files.
This is also important for making score files easier to
edit in Inkscape.
TODO 2: In Moritz, remove the code that creates
system markers and the invisible layer.