InputControls improvements
Removed all references to noteOnKey option
(see the updated comment in
svgScoreExtensions.htm).
Added support for the minVelocity attribute
which is now written by Moritz.
Moritz can now also write a value of "off" for
the noteOnVel, noteOff, pressure, pitchwheel,
modulation and speedOptions. This value
always cancels an existing performer option,
and leads to the values in the score being
performed as written.
If an option is undefined in the score, the
existing option value is used (which may be
"off").
While constructing the trks in the Assistant
Performer, the inputControls objects now
always have explicit "off" values for the
options not in use. These "off" values are
removed immediately before pushing the
inputControls values to each trk, so that
runtime memory use is minimized.