Only track InputControls now contain "off" values
"off" values are no longer created by the
InputControls constructor unless specifically
set in the score. A complete set of "off" values
is however maintained at the track level while
constructing Seqs.
Minimum InputControls continue to be
constructed in and sent to Workers from the Seq
constructor.
This allows options to be turned off by explicit
"off" values written in the score.