Bug Fixes in the speed option
The speed option has now been tested in all
the scores, and works without crashing, even
at high speeds.
Score.changeSpeed() has been moved to a later
point in the construction of the final timeObjects
so as to remove the possibility of rounding errors.
Account is now taken of msDurations that are
shorter than the duration between slider moments.