Bug Fixes etc. in Controls.js (!)
Don't call setMIDIDevices() inside the
onMIDIDeviceStateChange handler.
This was done previously, by turning the handler
off during the call, but turning it on again puts
the WebMIDIAPI shim (i.e. Firefox) into an endless
loop.
setMIDIDevices() is now called in beginRuntime(),
when the devices are first needed. There is no
reason to set them earlier!
Updated to the latest WebMIDIAPI shim, and
Firefox is now working too!