Working on the carrying forward of trkOptions.
Pedal, velocity and trkOff trkOptions are now
carried forward from one inputChord to the next.
TODO: The continuous controller options should
be attached to the _instant_ at the beginning of
the inputChord, not to individual notes.
Continuous controller options should be sent once,
immediately when the instant is first encountered,
(before any of the seqs or trkOffs are sent).
They are not really trkOptions at all. Maybe rename
trkOptions -> perfOptions?