Updated to read new ccSettings format.
Continuous controller options are now _separate_
from the trkOptions. They are attached to
inputChordDefs, not to individual notes, and can
only exist once at any particular instant in the score.
The changes seems to run correctly, but this needs
more testing to see that the continuous controller
messages are being sent and interpreted correctly
during live performances.
TODO: update Moritz so that it _writes_ the new format.