Added trkOptions.speed (!!)
This is currently tested and working.
Trks can now be made to perform at different speeds.
This allows, for example, trks to be defined that play
for much longer than their notated duration.
Such trks can be used polyphonically in live
performances to provide background (such as the
Wind in Song Six) that must be longer than the
performed durations of other tracks.
Time becomes "three dimensional" -- as if trks are
really notated at right angles to the plane of the page.