Implemented "disabled" trkOptions
pedal, velocity and noteOff trkOptions can now be set
to "disabled" in the score, so that other defined options
can be overridden by an option that says "do nothing".
Such "disabled" score values are simply left undefined
in the trkOptions pushed by Seq into the TrackWorker.
The (simple) code responsible for dealing with this is in
Keyboard1.setSeqTrkOptions().