Update for three private commits
These collectively solve two bugs in assisted
performances:
Bug #1: performances of single trks ended with an
extra note. Fixed by separating the sending of
allSoundOff and allControllersOff messages.
Bug#2: When a trk is started while a trk is already
playing in the track, the performing trk's silent
messages are now sent before the new trk starts
performing.

commit #1
finalChordOffMoment changes

The finalChordOffMoment in a MidiChord is
moved to the following MidiObject for default
performances (as before). But it is now kept as
an attribute of its MidiChord, and restored at
the end of a Trk (in Seq.getMoments()).

commit #2
TrackWorkers now created only when needed.

If a track is disabled for an assisted performance,
then no trackWorker is created for that track.
This is more efficient, and makes debugging easier.

commit #3
Improved TrackWorker behaviour

Replaced "trkCompleted" message by
"trkStopped" message. TrackWorkers send
this when the trk they are sending is stopped
prematurely. Keyboard1 then sends the
remaining silent messages (noteOffs and
controller messages) before sending the
messages for the following trk.
Also: The Keyboard1.resetChannel() function
has been replaced by a doAllSoundOff()
function that is called at the end of a live
performance, and a doAllControllersOff()
function that is called at the beginning of a
performance.