Changed to the Resident SoundFont Synth
The Resident SoundFont Synth is initially a
renamed copy of Sf2Synth1, located on the
same web site as the host, but in a directory
outside the host's directory.
The Resident SoundFont Synth may change
and develop as it is used in the Assistant
Performer.
Sf2Synth1 is now archived at its present state,
pending discussions about the Web MIDI API.